+++ src/utils/tcpsocket.hpp
@@ -13,7 +13,7 @@ class TCPSocket
 public:
     TCPSocket();
 
-    void connect(const char* addr, int port);
+    void connect(const char* addr, unsigned int port);
     void disconnect();
     bool isConnected();
 
@@ -23,17 +23,17 @@ public:
     std::string receive_blocking();
 
 private:
-    void setAddressAndPort(const char* addr, int port);
+    void setAddressAndPort(const char* addr, unsigned int port);
 #if defined __linux__ || defined __APPLE__
     void setPOSIXSocketAddress();
 #elif defined _WIN32
 #endif
 
     std::string mAddress;
-    int mPort;
+    unsigned int mPort;
     int mSocketId;
     bool mIsConnected;
-    const int mReceivingBufferSize;
+    const unsigned long int mReceivingBufferSize;
     sockaddr_in mServerAddress;
 };
